backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.9.0
boto3==1.35.0
botocore==1.35.0
python-dotenv==1.1.1
Pillow==11.3.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
cryptography==43.0.0
razorpay==1.3.0
django-storages==1.12.3
sentry-sdk[django]==2.19.0
django-redis==5.4.0
celery==5.3.4
redis==5.0.1